rename time span selector method + create FinalMerger() to get more modular code
